+++ app/src/main/java/com/renato/sofascoreacademy4/ScrollingActivity.kt
@@ -28,6 +28,7 @@ class ScrollingActivity : AppCompatActivity() {
             transformations(CircleCropTransformation())
             placeholder(R.drawable.elon_musk)
         }
+
         setText()
 
     }
