+++ app/src/main/res/values-night/themes.xml
@@ -44,7 +44,7 @@
         <item name="colorPrimaryInverse">@color/md_theme_dark_inversePrimary</item>
 
         <item name="android:statusBarColor" tools:targetApi="l">#7d6115</item>
-
+        <item name="android:textColor">#000000</item>
         <item name="android:fontFamily">@font/dosis</item>
     </style>
